Got wireless camera working with DirectShow by integrating AForge.Video.DirectShow with DShowNET and DirectX.Capture, modifying the filter graphs, and configuring the sample grabber to accept YUV inputs by writing a YUV to RGB24 Color Space Converter in code (change bool YUV to true or false in the sample grabber to activate/deactivate, must be true for our wireless camera and false for the lifecam).
I still need to complete IAMStreamConfig integration with AForge.Video.DirectShow in order to change the video input stream's resolution. Currently, the native 640x480 resolution of the wireless camera coupled with the motion detector turned on and tracking objects brings the computer to its knees as it has to compare up to 30 640x480 RGB24 bitmaps/second (each converted in real time from YUV Color Space), along with handling everything else on top of that. I need to reduce the input stream to a much more manageable 320x240 resolution for the sample grabber.